backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.0
anthropic==0.18.1
requests==2.31.0
beautifulsoup4==4.12.3
wikipedia-api==0.6.0
newsapi-python==0.2.7
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
